+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [2020.03.07]
 
   - Language switcher, refs #61, #62 and #63